+++ example/lib/main.dart
@@ -226,6 +226,30 @@ class _MyHomePageState extends State<MyHomePage> {
                 _selectFile();
               },
             ),
+            FlatButton(
+              child: Text("delete Document"),
+              onPressed: () {
+                _deleteFiles();
+              },
+            ),
+            FlatButton(
+              child: Text("Fetch Documents"),
+              onPressed: () {
+                _fetchFiles();
+              },
+            ),
+            FlatButton(
+              child: Text("Edit Document With file"),
+              onPressed: () {
+                _editFile(false);
+              },
+            ),
+            FlatButton(
+              child: Text("Edit Document Without file"),
+              onPressed: () {
+                _editFile(true);
+              },
+            ),
           ],
         ),
       ),
@@ -287,7 +311,9 @@ class _MyHomePageState extends State<MyHomePage> {
           copiedFile.path,
           "b55306bc-20d0-4ee6-adb1-d3307c308502",
           "2ac7d50d-da40-41a8-b84d-3a87c0fb9e4a",
-          "{\"displayName\": \"document name\",\"description\": \"description\",\"createDate\":\"2020-09-02T07:30:47\",\"visibility\":\"PUBLIC\",\"typeId\": \"3c3b5ae4-955c-4b9d-8bd9-3d564767a2e8\"}",
+          "{\"displayName\": \"document name\",\"description\": \"description\",\"createDate\":\"2020-09-02T07:30:47"
+              "\",\"visibility\":\"PUBLIC\",\"typeId\": \"3c3b5ae4-955c-4b9d-8bd9-3d564767a2e8\",\"mimeType\": "
+              "\"image/jpg\"}",
           (json) {
             Log.d(json, tag);
           },
@@ -296,5 +322,69 @@ class _MyHomePageState extends State<MyHomePage> {
         });
       }
     }
+  }
+
+  Future<void> _editFile(bool dataOnly) async {
+    var selectedFile = dataOnly ? null : await ImagePicker.pickImage(source: ImageSource.gallery).catchError((error) {
+      Log.error("selectImage", error: error);
+    });
+
+    var copiedFile;
+
+    if (selectedFile != null) {
+      var emptyFile = await BaseFileUtils.getLocalFile(
+        "imagesDir",
+        '${DateTime
+            .now()
+            .millisecondsSinceEpoch
+            .toString()}.jpg',
+      );
+
+      copiedFile = await selectedFile.copy(emptyFile.path).catchError((error) {
+        Log.error("selectImage copy", error: error);
+      });
+    }
+      // if (copiedFile != null && await copiedFile.exists()) {
+        NetworkManager(token).updateFileFuture(
+          pathFile: copiedFile?.path,
+          companyId: "b55306bc-20d0-4ee6-adb1-d3307c308502",
+          workspaceId: "2ac7d50d-da40-41a8-b84d-3a87c0fb9e4a",
+          documentId: "b9c8bfc6-9e94-4fef-9c35-725958e21090",
+          data: "{\"displayName\": \"document name1\",\"description\": \"description1\",\"createDate\":"
+              "\"2020-09-02T07:30:47"
+              "\",\"visibility\":\"PUBLIC\",\"typeId\": \"3c3b5ae4-955c-4b9d-8bd9-3d564767a2e8\",\"keepResource\": $dataOnly}",
+              //,\"mimeType\": \"image/jpg\"
+          handlePositiveResultBody: (json) {
+            Log.d(json, tag);
+          },
+        ).catchError((e) {
+          Log.error("uploadImageFuture", error: e);
+        });
+      // }
+    // }
+  }
+
+  Future<void> _fetchFiles() async {
+    NetworkManager(token)
+        .getFile(
+      "b55306bc-20d0-4ee6-adb1-d3307c308502",
+      "2ac7d50d-da40-41a8-b84d-3a87c0fb9e4a",
+    )
+        .catchError((e) {
+      Log.error("uploadImageFuture", error: e);
+    });
+  }
+
+  Future<void> _deleteFiles() async {
+    NetworkManager(token)
+        .deleteFile(
+      "b55306bc-20d0-4ee6-adb1-d3307c308502",
+      "2ac7d50d-da40-41a8-b84d-3a87c0fb9e4a",
+      "c064f29e-1a64-4777-9acc-c0fcb9a68e55",
+      (_) {},
+    )
+        .catchError((e) {
+      Log.error("uploadImageFuture", error: e);
+    });
   }
 }